RXV2 0.9.756-preflight-review — Pre-flight review of every firmware change since the last flown build (0.9.728), the night before Malcolm's next test flight.
Pre-flight review of every firmware change since the last flown build
(0.9.728), the night before Malcolm's next test flight. Verdict:
nothing in that diff can touch control in the air - every changed path
runs on the ground and is gated off when armed or under a live
transmitter link. It found three things worth fixing anyway. 1) The
tail-waggle fix from 0.9.729 never worked in the real landing order:
the rotor-speed sample sat inside a block that only runs after the
safety switch, by which time the motor already reads zero, so the
20-second coast wait never engaged and the tail still waggled on a
turning head. The rotor is now watched every tick - one comparison. 2)
The stall that switches the radios off after arming waited up to 1.5 s
for the transmitter to acknowledge its warning; a transmitter busy with
its safety-switch sound could push that stall to over two seconds after
the arm switch, into the spool-up for a pilot quick on the motor
switch. That wait is now capped at half a second, so the stall lands
within 0.8 s - close to the +0.3 s every flight so far has flown with.
3) The 20-second WiFi hold after a Bluetooth update never held: its
timer was compared unsigned against a time in the future and wrapped,
so the join started 200 ms after boot as before. Also: the new
bank-count setting is now refused under a live transmitter link, like
every other setting saved to flash. Firmware only.

Firmware only; pages unchanged (receivers skip the fs flash).

+++ RXV2/src/Network.h
@@ -413,7 +413,12 @@ inline void netStep() {
                 onWifiConnected();
                 break;
             }
-            if ((uint32_t)(millis() - netStateStart) >= WIFI_CONNECT_MS) {
+            // SIGNED: after a Bluetooth update netStateStart is set to the END of
+            // the 20 s STA hold (a time in the future), and the unsigned form
+            // wrapped to ~4e9 ms, "timed out" on the first pass, and overwrote
+            // the hold with a re-begin 200 ms later - the hold never held
+            // (found by the 0.9.756 pre-flight review). Negative = not yet.
+            if ((int32_t)(millis() - netStateStart) >= (int32_t)WIFI_CONNECT_MS) {
                 // If a phone is on our softAP, DON'T churn the STA side now — a
                 // disconnect+scan hops channels and boots that phone off mid-
                 // session. Hold the retry until they're done (self-heal resumes
